added edit-post views and routes

+++ controllers/api/postRoutes.js
@@ -22,35 +22,55 @@ router.post('/', withAuth, async (req, res) => {
 
 router.put('/:id', withAuth, async (req, res) => {
   try {
-    const postData = await Post.update(req.body, {
-      where: {
-        id: req.params.id,
-      },
-    });
+    userId = parseInt(req.session.user_id);
+    if (req.params.id && req.body) {
+      const postData = await Post.update(
+        { 
+          title: req.body.title,
+          content: req.body.content
+        },
+        {
+        where: {
+          id: req.params.id,
+          user_id: userId
+        },
+      });
+      if (!postData) {
+        res.status(404).json({ message: 'No post found with this id!' });
+        return;
+      };
 
+      res.status(200).json(postData);
+    } else {
+      res.status(404).json({ message: 'No post found with this id!' });
+    };
+    
     } catch (err) {
         res.status(500).json(err);    
-    }
-})
+    };
+});
 
 router.delete('/:id', withAuth, async (req, res) => {
   try {
-    const postData = await Post.destroy({
-      where: {
-        id: req.params.id,
-        user_id: req.session.user_id,
-      },
-    });
-
-    if (!postData) {
+    if (req.params.id) {
+      const postData = await Post.destroy({
+        where: {
+          id: req.params.id,
+          user_id: parseInt(req.session.user_id),
+        },
+      });
+      
+      if (!postData) {
+        res.status(404).json({ message: 'No post found with this id!' });
+        return;
+      };
+      res.status(200).json(postData);
+    } else {
       res.status(404).json({ message: 'No post found with this id!' });
-      return;
-    }
-
-    res.status(200).json(postData);
+    };
   } catch (err) {
     res.status(500).json(err);
-  }
+  };
 });
 
 module.exports = router;
